Cleaned up javascript and fixed Firefox compatibility issues
Removed parallel.js (which contained the Game of Life implemented
using gpu.js, but isn't integrated into the pipeline). Some
initialization which was in parallel.js was moved to ui.js.

Fixed several javascript errors which aren't reported when executing
on Chrome, but cause errors when executing on Firefox, mostly undefined
references.